@@ -3,80 +3,38 @@
 namespace App\Livewire\Admin\Users;
 
 use App\Enums\Can;
+use App\Helpers\Table\Header;
 use App\Models\{Permission, User};
+use App\Traits\Livewire\HasTable;
 use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\{Builder, Collection};
-use Illuminate\Pagination\LengthAwarePaginator;
-use Livewire\Attributes\{Computed, On, Rule};
+use Livewire\Attributes\{On, Rule};
 use Livewire\{Component, WithPagination};
 
 class Index extends Component
 {
     use WithPagination;
+    use HasTable;
 
-    protected $paginationTheme = 'tailwind';
+    public bool $filtersVisible = false;
 
-    public ?string $search = null;
-
-    public bool $search_trash = false;
+    public bool $searchTrash = false;
 
     #[Rule(['exists:permissions,id'])]
-    public array $search_permissions = [];
+    public array $searchPermissions = [];
 
     public Collection $permissionsToSearch;
 
-    public string $sortDirection = 'asc';
-
-    public string $sortColumnBy = 'id';
-
-    public int $perPage = 10;
-
-    #[Computed]
-    public function users(): LengthAwarePaginator
-    {
-        $this->validate();
-
-        return User::query()
-            ->with('permissions')
-            ->when($this->search, function (Builder $query) {
-                $query->whereRaw('lower(name) like ?', ['%' . strtolower($this->search) . '%'])
-                    ->orWhere(
-                        'email',
-                        'like',
-                        '%' . strtolower($this->search) . '%'
-                    );
-
-            })->when($this->search_permissions, function (Builder $query) {
-                $query->whereHas('permissions', function (Builder $query) {
-                    $query->whereIn('id', $this->search_permissions);
-                });
-
-            })->when($this->search_trash, function (Builder $query) {
-                $query->onlyTrashed();
-
-            })->orderBy($this->sortColumnBy, $this->sortDirection)
-            ->paginate($this->perPage);
-    }
-
-    #[Computed]
-    public function headers(): array
-    {
-        return [
-            ['key' => 'id', 'label' => '#'],
-            ['key' => 'name', 'label' => 'Name'],
-            ['key' => 'email', 'label' => 'Email'],
-            ['key' => 'permissions', 'label' => 'Permissions'],
-        ];
-    }
-
-    #[Computed]
     public function filterPermissions(?string $value = null): void
     {
+        $selectedOption = Permission::whereIn('id', $this->searchPermissions)->get();
+
         $this->permissionsToSearch = Permission::query()
             ->when($value, function (Builder $query) use ($value) {
                 $query->where('key', 'like', "%$value%");
             })->orderBy('key')
-            ->get();
+            ->get()
+            ->merge($selectedOption);
     }
 
     public function mount(): void
@@ -97,10 +55,40 @@ class Index extends Component
         $this->resetPage();
     }
 
-    public function sortBy(string $column, string $direction): void
+    public function toggleFilters(): void
     {
-        $this->sortColumnBy  = $column;
-        $this->sortDirection = $direction;
+        $this->filtersVisible = !$this->filtersVisible;
+    }
+
+    /**
+     * @return Header[]
+     */
+    public function tableHeaders(): array
+    {
+        return [
+            Header::make('id', '#'),
+            Header::make('name', 'Name'),
+            Header::make('email', 'Email'),
+            Header::make('permissions', 'Permissions'),
+        ];
+    }
+
+    public function searchColumns(): array
+    {
+        return ['name', 'email'];
+    }
+
+    public function query(): Builder
+    {
+        return User::query()
+            ->with('permissions')
+            ->when($this->searchPermissions, function (Builder $query) {
+                $query->whereHas('permissions', function (Builder $query) {
+                    $query->whereIn('id', $this->searchPermissions);
+                });
+            })->when($this->searchTrash, function (Builder $query) {
+                $query->onlyTrashed();
+            });
     }
 
     public function destroy(int $id): void
