@@ -27,7 +27,7 @@ class UserFactory extends Factory
             'email'             => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
             'validation_code'   => null,
-            'password'          => static::$password ??= Hash::make('password'),
+            'password'          => Hash::make('password'),
             'remember_token'    => Str::random(10),
         ];
     }
