@@ -55,7 +55,8 @@
                     </x-list-item>
                 @endif
 
-                <x-menu-item title="Home" icon="o-home" link="/" />
+                <x-menu-item title="Home" icon="o-home" :link="route('dashboard')" />
+                <x-menu-item title="Customers" icon="o-building-storefront" :link="route('customers')" />
 
                 @can(\App\Enums\Can::BE_AN_ADMIN->value)
                     <x-menu-sub title="Admin" icon="o-lock-closed">
