@@ -35,8 +35,8 @@ it("let's create a livewire component to list as users in the page", function ()
 
     $livewire = Livewire::test(Admin\Users\Index::class);
 
-    $livewire->assertSet('users', function ($users) {
-        expect($users)
+    $livewire->assertSet('items', function ($items) {
+        expect($items)
             ->toBeInstanceOf(LengthAwarePaginator::class)
             ->toHaveCount(10);
 
@@ -55,10 +55,10 @@ it('check the table format', function () {
 
     Livewire::test(Admin\Users\Index::class)
         ->assertSet('headers', [
-            ['key' => 'id', 'label' => '#'],
-            ['key' => 'name', 'label' => 'Name'],
-            ['key' => 'email', 'label' => 'Email'],
-            ['key' => 'permissions', 'label' => 'Permissions'],
+            ['key' => 'id', 'label' => '#', 'sortColumnBy' => 'id', 'sortDirection' => 'asc'],
+            ['key' => 'name', 'label' => 'Name', 'sortColumnBy' => 'id', 'sortDirection' => 'asc'],
+            ['key' => 'email', 'label' => 'Email', 'sortColumnBy' => 'id', 'sortDirection' => 'asc'],
+            ['key' => 'permissions', 'label' => 'Permissions', 'sortColumnBy' => 'id', 'sortDirection' => 'asc'],
         ]);
 });
 
@@ -66,27 +66,28 @@ it('should be able to filter by name and email', function () {
     /** @var User $admin */
     $admin = User::factory()->admin()->create(['name' => 'Is Admin', 'email' => 'is_dmin@mail.com']);
     actingAs($admin);
+
     User::factory()->create(['name' => 'Is User', 'email' => 'any_email@mail.com']);
 
     Livewire::test(Admin\Users\Index::class)
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toBeInstanceOf(LengthAwarePaginator::class)
                 ->toHaveCount(2);
 
             return true;
         })
         ->set('search', 'user')
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toHaveCount(1)
                 ->first()->name->toBe('Is User');
 
             return true;
         })
         ->set('search', 'any')
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toHaveCount(1)
                 ->first()->name->toBe('Is User');
 
@@ -98,20 +99,21 @@ it('should be able to filter by permission.key', function () {
     /** @var User $admin */
     $admin = User::factory()->admin()->create(['name' => 'Is Admin', 'email' => 'is_dmin@mail.com']);
     actingAs($admin);
+
     User::factory()->create(['name' => 'Is User', 'email' => 'any_email@mail.com']);
     $permission = Permission::where('key', '=', Can::BE_AN_ADMIN)->first();
 
     Livewire::test(Admin\Users\Index::class)
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toBeInstanceOf(LengthAwarePaginator::class)
                 ->toHaveCount(2);
 
             return true;
         })
-        ->set('search_permissions', [$permission->id])
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->set('searchPermissions', [$permission->id])
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toHaveCount(1)
                 ->first()->name->toBe('Is Admin');
 
@@ -122,20 +124,21 @@ it('should be able to filter by permission.key', function () {
 it('should be able to list deleted users', function () {
     /** @var User $admin */
     $admin = User::factory()->admin()->create(['name' => 'Is Admin', 'email' => 'is_dmin@mail.com']);
-    User::factory(2)->create(['deleted_at' => now()]);
     actingAs($admin);
 
+    User::factory(2)->create(['deleted_at' => now()]);
+
     Livewire::test(Admin\Users\Index::class)
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->toBeInstanceOf(LengthAwarePaginator::class)
                 ->toHaveCount(1);
 
             return true;
         })
-        ->set('search_trash', true)
-        ->assertSet('users', function ($users) {
-            expect($users) ->toHaveCount(2);
+        ->set('searchTrash', true)
+        ->assertSet('items', function ($items) {
+            expect($items)->toHaveCount(2);
 
             return true;
         });
@@ -144,27 +147,28 @@ it('should be able to list deleted users', function () {
 it('should be able to sort by id, name and email', function () {
     /** @var User $admin */
     $admin = User::factory()->admin()->create(['name' => 'Is Admin', 'email' => 'is_dmin@mail.com']);
-    User::factory()->create(['name' => 'Is User', 'email' => 'any_email@mail.com']);
     actingAs($admin);
+
+    User::factory()->create(['name' => 'Is User', 'email' => 'any_email@mail.com']);
 
     // ASC => Is Admin, Is User
     // DESC => Is User, Is Admin
     Livewire::test(Admin\Users\Index::class)
         ->set('sortDirection', 'asc')
         ->set('sortColumnBy', 'name')
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->first()->name->toBe('Is Admin')
-                ->and($users)->last()->name->toBe('Is User');
+                ->and($items)->last()->name->toBe('Is User');
 
             return true;
         })
         ->set('sortDirection', 'desc')
         ->set('sortColumnBy', 'name')
-        ->assertSet('users', function ($users) {
-            expect($users)
+        ->assertSet('items', function ($items) {
+            expect($items)
                 ->first()->name->toBe('Is User')
-                ->and($users)->last()->name->toBe('Is Admin');
+                ->and($items)->last()->name->toBe('Is Admin');
 
             return true;
         });
@@ -173,20 +177,21 @@ it('should be able to sort by id, name and email', function () {
 it("should be able to paginate the result", function () {
     /** @var User $user */
     $user = User::factory()->admin()->create();
-    User::factory(30)->create();
     actingAs($user);
 
+    User::factory(30)->create();
+
     Livewire::test(Admin\Users\Index::class)
-    ->assertSet('users', function ($users) {
-        expect($users)
+    ->assertSet('items', function ($items) {
+        expect($items)
             ->toBeInstanceOf(LengthAwarePaginator::class)
             ->toHaveCount(10);
 
         return true;
     })
     ->set('perPage', 15)
-    ->assertSet('users', function ($users) {
-        expect($users)
+    ->assertSet('items', function ($items) {
+        expect($items)
             ->toBeInstanceOf(LengthAwarePaginator::class)
             ->toHaveCount(15);
 
